rename phys_velocity
