+++ src/phys_velocity.ts
